+++ src/app/data/models.ts
@@ -180,6 +180,7 @@ export interface Settings {
   offlineEvents: string[]; // List of all offline event ids
   lastGeoAlert: number | undefined; // Last geolocation permission request alert
   lastAboutAlert: number | undefined; // Last time the user got a message about the selected event
+  lastLongEvents: number | undefined; // Last time the user got a message about long events
 }