+++ src/app/map/map.component.ts
@@ -107,6 +107,10 @@ export class MapComponent implements OnInit, OnDestroy {
     return this._points;
   }
 
+  public async capture(): Promise<string | undefined> {
+    return await this.mapResult?.capture();
+  }
+
   private async fixGPSAndUpdate() {
     this.selectedPoint = undefined;
     let foundPoints = 0;
@@ -282,6 +286,7 @@ export class MapComponent implements OnInit, OnDestroy {
       height: 0,
       defaultPinSize: pinSize,
       pins: [],
+      backgroundColor: this.ui.darkMode() ? 0x111111 : 0x999999,
       compass: { uuid: 'compass', x: compassPt.x, z: compassPt.y, color: 'compass', size: pinSize, label: '' },
       pinClicked: this.pinClicked.bind(this),
     }