+++ changelog.md
@@ -1,5 +1,12 @@
 ## Changes
 
+## 2.43
+
+- Maps can be shared (eg dropped pins)
+- Fix camps and art when text zoom is changed
+- Only message about long events once per day
+- Swipe back now working on now > event page
+
 ## 2.42
 
 - Additional labelling added to the Burning Man map